requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
dj-rest-auth
django-cors-headers
drf-yasg
django-private-storage